0.10.2

# -----BEGIN PGP SIGNATURE-----
#
# iQIzBAABCAAdFiEEzT/p4dnsSMryVjng3N1RYVzHNM0FAmZjVYMACgkQ3N1RYVzH
# NM0dfw//SNRdMBN+k0W7WXFbMOOw6FYDENxztuGwKtvK1owoUKedCwvLMvAVZcgB
# Rt2hwyCS2oQ+2GMRtG9vj4vGXlXmvI7BDK8JLoG7ZAVQX8kTSVqwrDaf+2dG8SOL
# lRl8U2SR6Flky3ZprLLv7xdYboC89lz0GK0MtZJOPw/WpPKYQXbMtCJ77wfAqptx
# M1RYyE3Z9f7IVubstExBifRC//jZ5TwtspBlDGweKVS2GN7Fuufqd6mCz9N/LdYX
# QW0+KpMRnOblJsUPlMMfkj1tgZ9GVm3w2gyrOdMQIPDxlLp4r9OSXbHnKaVnG8xS
# 5TM1px5t91wI3Kr5q9LKFrR5IVYTE27einbFDfp825xwX/Is/FGaM1OusRQSVq2A
# mVZpXQDURIcWo6WZsm+VBVvec2tcDYjF1/47LP/4tmrVficIhAOfrdNyATis7d3D
# Tpw4gFA6o1VHRjsh33Y8twOlz0OZuvR+jKQA/ZInbj2u9pDF35T6S0N//CERT5n7
# KsAQOuNohp3MelDGL7kDgT/JC59JyUAH9/2TldLAu1UfX3M7h7oFQsfJptcOk2Il
# yI9HRJ35Qmy5ZMY0JmA5VdT8pbINEsFvpfJ/1uNxBs/fTo+aWTFJa2TgQCyCZ7MU
# R3ehdNumNhNONrXTcyK6nCoUIW15ET4rpGciRRyeD+WVTUmq7MY=
# =ShMO
# -----END PGP SIGNATURE-----
# gpg: Signature made Fri Jun  7 19:46:27 2024 WEST
# gpg:                using RSA key CD3FE9E1D9EC48CAF25639E0DCDD51615CC734CD
# gpg: Good signature from "João Magalhães <joamag@gmail.com>" [unknown]
# gpg: WARNING: This key is not certified with a trusted signature!
# gpg:          There is no indication that the signature belongs to the owner.
# Primary key fingerprint: CD3F E9E1 D9EC 48CA F256  39E0 DCDD 5161 5CC7 34CD

+++ frontends/sdl/src/sdl.rs
@@ -35,14 +35,18 @@ impl SdlSystem {
         let audio_subsystem = sdl.audio().unwrap();
         let event_pump = sdl.event_pump().unwrap();
 
-        // initialized the fonts context to be used
-        // in the loading of fonts
+        // initializes the fonts context to be used
+        // in the loading of fonts (supports TTF fonts)
         let ttf_context = sdl2::ttf::init().unwrap();
 
         // creates the system window that is going to be used to
         // show the emulator and sets it to the central are o screen
         let window = video_subsystem
-            .window(title, scale as u32 * width, scale as u32 * height)
+            .window(
+                title,
+                (scale * width as f32) as u32,
+                (scale * height as f32) as u32,
+            )
             .resizable()
             .position_centered()
             .opengl()
